Rename rpc to fsrpc
