+++ fsrpc/README.md
@@ -1,4 +1,4 @@
-## RPC
+## fsrpc
 
 This package implements RPC interfaces that connected clients can call in KBFS,
 to do certain operations, such as listing files.
